move root_disk() from fs/btrfs to system/osi #1320
Not directly related to btrfs so moving to more general location.
